backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.6.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
